Move layout related components under `layout/`
