--- admin/app/components/solidus_admin/feedback/component.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-class SolidusAdmin::Feedback::Component < SolidusAdmin::BaseComponent
-end
