+++ admin/app/components/solidus_admin/layout/feedback/component.html.erb
