+++ admin/app/components/solidus_admin/layout/feedback/component.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class SolidusAdmin::Layout::Feedback::Component < SolidusAdmin::BaseComponent
+end
