+++ admin/app/components/solidus_admin/layout/navigation/account/component.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+# Account navigation
+class SolidusAdmin::Layout::Navigation::Account::Component < SolidusAdmin::BaseComponent
+  # @param user_label [String]
+  # @param account_path [String]
+  # @param logout_path [String]
+  # @param logout_method [Symbol]
+  def initialize(user_label:, account_path:, logout_path:, logout_method:)
+    @user_label = user_label
+    @account_path = account_path
+    @logout_path = logout_path
+    @logout_method = logout_method
+  end
+end
