+++ admin/app/components/solidus_admin/layout/navigation/account/component.yml
