+++ admin/app/components/solidus_admin/layout/navigation/component.yml
