--- admin/spec/components/previews/solidus_admin/feedback/component_preview.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-# @component "ui/feedback"
-class SolidusAdmin::Feedback::ComponentPreview < ViewComponent::Preview
-  include SolidusAdmin::Preview
-
-  def overview
-    render_with_template
-  end
-end
