+++ admin/spec/components/previews/solidus_admin/layout/feedback/component_preview.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class SolidusAdmin::Layout::Feedback::ComponentPreview < ViewComponent::Preview
+  include SolidusAdmin::Preview
+
+  def overview
+    render_with_template
+  end
+end
