+++ admin/spec/components/previews/solidus_admin/layout/navigation/item/component_preview/overview.html.erb
