--- admin/spec/components/solidus_admin/skip_link/component_spec.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require "spec_helper"
-
-RSpec.describe SolidusAdmin::SkipLink::Component, type: :component do
-  it "renders the overview preview" do
-    render_preview(:overview)
-  end
-end
